+++ NEWS.md
@@ -1,7 +1,8 @@
 
-# frac 0.1.1.9000  2024-12-12
+# frac 0.1.1.9001  2024-12-12
 
-* Modernise
+* [9001] Add base R julia calculation
+* [9000] Modernise
 
 # frac 0.1.1  2020-07-19